removed test from serde_utils_test to utils_test
